refactor(feedback-source): rename reconcile module to mapping-reconciliation
ENG-2058 (#8798) already adds a lib/feedback-source/reconcile.ts for a different
concern - retrying Hub feedback-record creates after a 409. Two unrelated modules on
one path would land as an add/add conflict where either resolution silently drops one
of them, so this one takes the name that says what it reconciles.
